Implemented suggested Directory Layout. Renamed a handful of classes. Set up automatic binding to a Main so we don't need to run from specific file Updated build script to be more dynamic with intentions Updated ReadMe to explain intention for directory layout.
