+++ src/main/java/com/bookclub/model/User.java
@@ -1,4 +1,4 @@
-package com.example.organizer;
+package com.bookclub.model;
 
 public class User {
     public User() {
